requirements.txt
django>4.2.0,<5.1.1
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2
pillow
algoliasearch-django>=3.0,<4.0